สำรอง

+++ Verify/php_CheckVerify.php
@@ -4,11 +4,14 @@ include_once('../database/connectDB.php');
 
 if(isset($_POST['submit'])){
     if($_SESSION['numberOTP']==$_POST['otp']){
+
         $sql = "UPDATE `user` SET `status` = '1' WHERE `user`.`id_user` = '".$_SESSION['id_user']."';";
         $result = $conn->query($sql) or die($conn->error);
 
-        $sql_delete = "DELETE FROM `user` WHERE `email` = '".$_SESSION['email_user']."' AND status <> 1 ";
-        $result_delete = $conn->query($sql_delete);
+        if($result){
+            $sql_delete = "DELETE FROM `user` WHERE `email` = '".$_SESSION['email_user']."' AND status <> 1 ";
+            $result_delete = $conn->query($sql_delete);
+        }
 
         $_SESSION['id_user'] = NULL;
         $_SESSION['email_user'] = NULL;
